ci: fine-tune commit creation workflow
If the title of the pull request adheres to the conventional commit format, an empty commit will be created using the title as the commit header. Additional details such as the commit body and footer can also be added through the PR body. Multiple commits may be created by adding semicolons to the title.