ci: change release trigger to pull request
The main branch is protected and requires a PR in order to merge. However, PRs require owner approval and the approved label. In order to avoid all of the overhead in bypassing these processes, a new label has been created that will be applied during the PR process. This label will trigger the release workflow.